+++ plugins/core/renderer_gl/src/services.c
@@ -25,7 +25,7 @@ register_services(const struct plugin_t* plugin)
 
     SERVICE_REGISTER5(plugin, "text_create",                    text_create_wrapper, uint32_t, char, uint32_t, float, float, const wchar_t*);
     SERVICE_REGISTER4(plugin, "text_destroy",                   text_destroy_wrapper, uint32_t, uint32_t, float, float, const wchar_t*);
-    SERVICE_REGISTER2(plugin, "text_set_centered",              text_set_centered_wrapper, void, uint32_t, intptr_t);
+    SERVICE_REGISTER2(plugin, "text_set_centered",              text_set_centered_wrapper, void, uint32_t, uint32_t);
     SERVICE_REGISTER1(plugin, "text_set_position",              text_set_position_wrapper, void, uint32_t);
     SERVICE_REGISTER1(plugin, "text_set_string",                text_set_string_wrapper, void, const wchar_t*);
     SERVICE_REGISTER1(plugin, "text_show",                      text_show_wrapper, void, uint32_t);
