+++ util/src/string.c
@@ -211,7 +211,7 @@ strtowcs(const char* str)
 {
     wchar_t* wcs;
     wchar_t* wcs_it;
-    intptr_t len = strlen(str);
+    uint32_t len = strlen(str);
     wcs = (wchar_t*)MALLOC((len + 1) * sizeof(wchar_t));
     for(wcs_it = wcs; *str; ++str)
         *wcs_it++ = (wchar_t)*str;
@@ -225,7 +225,7 @@ wcstostr(wchar_t* wcs)
 {
     char* str;
     char* str_it;
-    intptr_t len = wcslen(wcs);
+    uint32_t len = wcslen(wcs);
     str = (char*)MALLOC((len + 1) * sizeof(char));
     for(str_it = str; *wcs; ++wcs)
         *str_it++ = (char)*wcs;
